Move relevant tests from exp_domain_test to state_domain_test
